Decrease BitField size from u32 to u8
A bitfield can only be created from integral types, whose maximum size is 64 (which can easily fit inside a u8)

+++ src/encoding.rs
@@ -28,7 +28,7 @@ pub enum Encoding<'a> {
     Class,
     Sel,
     Unknown,
-    BitField(u32),
+    BitField(u8),
     Pointer(&'a Encoding<'a>),
     Array(usize, &'a Encoding<'a>),
     Struct(&'a str, &'a [Encoding<'a>]),
